backend/requirements.txt
Flask==3.1.1
Flask-Scss==0.5
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.6
PyMySQL==1.1.1
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.43
